Created a new class called UsersHandler which will centralize all operations related to the users collection. This is a utility class that will be used by the main "FirebaseDatabase" class
